geotiff tests: consolidate VRT tail cluster (#2437)
Fold the 12 VRT-residue files into existing vrt/ modules plus a new
vrt/test_parity.py:

- missing-sources trio (1843 internal entry point + STRICT env, 1860
  public default, 1799/2265 chunked policy) -> vrt/test_missing_sources.py
- srcrect-validation (1784), kwarg-drop (1685), tiled-validation (1862)
  -> vrt/test_validation.py
- lazy-chunks (1798), dask-vrt-kwargs (1795) -> vrt/test_window.py
- backend-parity (2321), finalization-parity (2162), backend-coverage
  (2026_05_11) -> new vrt/test_parity.py

Update release_gate_geotiff.rst rows to cite the consolidated homes so
the checklist-parity gate stays green. CLUSTER_AUDIT_VRT_TAIL.md maps
every old file::test to its new home and is deleted before merge.
